Discombob changes (QoL)
After using the discombob the time of the stun will count down in chat to give user feedback on how long the stun lasts. Everything SHOULD work as intended, it did in my tests. Feel free to test further and let me know if it brings up issues i couldnt find.

Some things in there might also be placeholders for if i can figure out a smooth way to update terminal with the countdown as well. it might be a little messy please forgive me lmfao