Implementation of Protein Powder's "Crit" ability
This implementation makes it so, whenever a player hits an enemy with a shovel, they roll a die to see if they deal a crit or not.
We first check if they have maxed out Protein Powder. No point rolling a die if we're not gonna accept the value anyways.
After checking that they maxed out Protein Powder, we roll a die and see if the value provided is lower than the value configured for the probability of dealing a crit.

If you wish to make this "crit" ability be like "Press 'x' for the next hit to be a crit", you can just change the "TryToCritEnemy" logic to check if they have previously pressed 'x' and attach a cooldown to pressing 'x'.