Implementation of Player Health Upgrade (Stimpack)
Made a new upgrade called "Stimpack" which basically increases the player's health.
Sounds easy enough right? Just increase the max health of the player and be going with it.
WELL, apparently it's just one variable for health and it's the current one so I had to be creative.

Whenever the player buys the upgrade or increments the level, check if it's the local player that bought it to update the health.
Whenever a new round starts, each player's health gets updated if they have any entries in the dictionary associated to this upgrade (which stores the amount of health they will have). If so, put the respective value into their health and update the HUD.

This obviously won't work if any other mods that regen health as they all, and rightfully so, set the health value to the default which is 100. For mods that use DamagePlayer with negative value, we could use a patcher where it checks it's negative and if it's a full vanilla heal. If so, set it the player's additional health value from the dictionary. This is if we really want the health boost to be still applying after a heal (that isn't ours since we're adding the medkit)