Fixing the scanner ignoring enemies that are infront of the player
We were just ignoring enemies whenever the config option was false, leading to people not being able to scan enemies at all.

The intended behaviour is, if a node is behind a wall, we check if it's an enemy and if so we check if we can scan enemies through walls.
If alright, let the node appear when scanning. Otherwise, don't make the node appear.