Update Hunter mod to (a) work and (b) be more extensible/configurable
Currently, the Hunter mod is hard-coded to only supporting three levels, and an off-by-one bug means that at level 3 everything breaks on a key not found error:
```
KeyNotFoundException: The given key '3' was not present in the dictionary.
  at System.Collections.Generic.Dictionary`2[TKey,TValue].get_Item (TKey key) [0x0001e] in <787acc3c9a4c471ba7d971300105af24>:IL_001E
  at MoreShipUpgrades.Patches.Enemies.EnemyAIPatcher.SpawnSample (EnemyAI __instance) [0x00053] in
```

This change reworks how enemy tiers are stored and retreived to avoid this issue and also allow for a variable number of monster tiers to be configurable.