Implemented new ship upgrade
An upgrade labelled "Efficient Engines" which reduces the cost of moon routing by provided discount.
# Configurable Values
- Enable the upgrade
- Initial price of the upgrade
- Prices of the subsequent levels of the upgrade
- Initial discount applied to moon routing
- Incremental discount applied to moon routing on further levels
# Player Feedback
To show the player that the discount is being used, we change the value displayed when routing to a moon before confirming said routing. Not so certain if this can break other UI elements in the terminal but so far it shows. The alternative would be changing the itemCost of all terminalNodes which route to a moon. :D