Implemented a player upgrade
# Climbing Gloves
Provides on boost on the player's speed at climbing ladders
# Configurable Entries
- Enable the upgrade
- Set it to individual or shared upgrade
- Initial price of the upgrade
- Prices of each subsequent level of the upgrade
- Initial climbing speed value applied on purchase
- Incremental climbing speed value applied on purchase

# Player Feedback
When the ``PlayerControllerB`` starts climbing ladder, the speed at which it goes up and down uses a variable called ``climbSpeed`` which affects how fast the model moves. We apply the same logic on all other player attribute upgrades to this one. Resets can happen if during play time the variable is reset to default.