Implemented "UniqueScrap" Item Progression mode
Whenever a give scrap item is sold, a portion of its scrap value is provided as a contribution towards buying its associated upgrade. The particularity of this mode is that the associations between scrap items and upgrades is random and not influenced by the players.
The item sold will be later displayed in the upgrade's page once sold for the first time.
The price of the upgrade will decrease accordingly when selling the scrap item
If new items are added after the initialization of the dictionary, they will be added later on save boot.